Fixed and restored operations break_loop, continue_loop, try_for_dict_keys. (requires allow_wse_execute_statement_blocks = 1 in module.ini) Lua iterator game.propInstI now work as exactly like try_for_prop_instances. Response for http request for server info now sends correct headers and opens on any browsers. (http://server_ip:server_port/) Added script wse_get_server_info. Bug fixes.

+++ WarbandLib/rgl_channel.h
@@ -10,7 +10,7 @@ namespace rgl
 	struct channel
 	{
 		float blend_progress;
-		int u1;
+		float u1;
 		unsigned __int64 enforced_bones;
 		unsigned __int64 rotated_bones;
 		float parameter;
